# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "APU EMS API",
    "version": "1.0.0",
    "description": "REST API for the APU Event Management System.\n\nThe backend owns the proposal workflow state machine and every authorisation decision. Clients send actions and render what comes back; no permission rule is evaluated in the browser.\n\n**Authentication.** `POST /auth/login` returns a short-lived access token and a longer-lived refresh token. Send the access token as `Authorization: Bearer <token>`. On 401 with code `token_expired`, call `POST /auth/refresh`.\n\n**Errors.** Every failure returns the same envelope: `{\"error\": {\"code\", \"message\", \"request_id\"}}`. `message` is safe to display; `request_id` correlates with server logs.\n\n**Rate limits.** 300 requests/minute overall, 10/minute on auth endpoints, keyed per authenticated user and falling back to client IP.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found"
              },
              "message": {
                "type": "string",
                "description": "Safe to show a user."
              },
              "request_id": {
                "type": "string",
                "description": "Correlates with server logs."
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "operationId": "uploads_upload_post",
        "summary": "Accept a base64 data URL and return {storageKey, url}.",
        "responses": {
          "201": {
            "description": "Success."
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Not allowed in the resource's current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/uploads/{key}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "operationId": "uploads_serve_get",
        "summary": "Serve a stored image.",
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "The key is matched against a strict pattern before touching the filesystem,\nso no caller-supplied path can escape the upload directory.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    }
  }
}